프로젝트 9 - NPC 완료
. 섹터 처리가 된 서버에 20만 개의 NPC 추가. NPC를 수용하기 위해 맵 크기를 2000x2000으로 확장.
. NPC는 랜덤 위치로 1초에 한 번씩 이동. 모든 NPC의 이동 처리를 할 경우 오버헤드가 심하므로, 플레이어 주위의 NPC만을 이동시키도록 최적화. 이 또한 섹터를 통해 검사함. 플레이어 주변의 섹터를 검사하여 시야 내의 NPC만 이동 처리를 하고, 이동 처리된 NPC의 시야를 다시 검사하여 시야 내의 플레이어에게 이동 패킷을 전송함.

+++ 9. NPC - Server/protocol.h
@@ -62,32 +62,32 @@ struct cs_packet_move : public packet
 
 struct cs_packet_logout : public packet
 {
-	unsigned short id;
+	unsigned int id;
 };
 
 // Server to Client
 
 struct sc_packet_login_confirm : public packet
 {
-	unsigned short id;
+	unsigned int id;
 };
 
 struct sc_packet_add_player : public packet
 {
-	unsigned short id;
+	unsigned int id;
 	Short2 coord;
 };
 
 struct sc_packet_object_info : public packet
 {
-	unsigned short id;
+	unsigned int id;
 	Short2 coord;
 	UINT moveTime;
 };
 
 struct sc_packet_exit_player : public packet
 {
-	unsigned short id;
+	unsigned int id;
 };
 
 #pragma pack(pop)
